Moved RemoteMessageHandler to remote_message_service file.
